+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [1.16.3] - 2025-08-22
+
+### Fixed
+
+- Fix double synchronization of transaction log + integrity checks, [PR-912](https://github.com/reductstore/reductstore/pull/912)
+
 ## [1.16.2] - 2025-08-22
 
 ### Fixed
@@ -1099,7 +1106,8 @@ reduct-rs: `ReductClient.url`, `ReductClient.token`, `ReductCientBuilder.try_bui
 
 - Initial release with basic HTTP API and FIFO bucket quota
 
-[Unreleased]: https://github.com/reductstore/reductstore/compare/v1.16.2...HEAD
+[Unreleased]: https://github.com/reductstore/reductstore/compare/v1.16.3...HEAD
+[1.16.3]: https://github.com/reductstore/reductstore/compare/v1.16.2...v1.16.3
 [1.16.2]: https://github.com/reductstore/reductstore/compare/v1.16.1...v1.16.2
 [1.16.1]: https://github.com/reductstore/reductstore/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/reductstore/reductstore/compare/v1.15.6...v1.16.0